@@ -1,24 +1,25 @@
 {
   "name": "reactjs-boilerplate",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
-    "loglevel": "^1.7.1",
-    "mobx": "^6.2.0",
-    "mobx-react": "^7.2.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.2",
-    "react-jss": "^10.6.0",
-    "react-localization": "^1.0.16",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.1"
+    "axios": "^0.27.2",
+    "babel-eslint": "^10.1.0",
+    "bootstrap": "^5.2.0",
+    "loglevel": "^1.8.0",
+    "mobx": "^6.6.1",
+    "mobx-react": "^7.5.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.5.0",
+    "react-dom": "^18.2.0",
+    "react-jss": "^10.9.2",
+    "react-localization": "^1.0.19",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -55,7 +56,7 @@
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "husky": "^7.0.2"
+    "husky": "^7.0.2",
     "prettier": "^2.2.1"
   }
 }
