@@ -1,18 +1,18 @@
 import React from 'react';
 import {
-  BrowserRouter, Switch, Route, Redirect,
+  BrowserRouter, Routes, Route, Navigate,
 } from 'react-router-dom';
 import { Home, PageError } from '../pages';
 
 export default () => (
   <BrowserRouter>
-    <Redirect
-      from="/"
-      to="/home"
-    />
-    <Switch>
-      <Route path="/home" component={Home} exact />
-      <Route component={PageError} />
-    </Switch>
+    <Routes>
+      <Route path="/home" element={<Home />} exact />
+      <Route element={<PageError />} />
+      <Route
+        path="*"
+        element={<Navigate to="/home" replace />}
+      />
+    </Routes>
   </BrowserRouter>
 );
